Refactor shrub into folder structure
