+++ src/Game/Blobs/blobQueen/types.ts
@@ -1,5 +1,5 @@
 import { Coordinates } from 'src/types';
-import { ShrubActor } from 'game/resources';
+import { ShrubActor } from 'game/resources/shub';
 import { BlobletActor } from '../bloblet';
 import { BlobLarvaActor } from '../blobLarva';
