+++ src/Game/Resources/shub/index.ts
@@ -1,0 +1,2 @@
+export * from './shub';
+export * from './types';
